% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-batch-correction-cydar.R
\name{batch_corrected_drop_na}
\alias{batch_corrected_drop_na}
\title{Title}
\usage{
batch_corrected_drop_na(batch_corrected_df)
}
\arguments{
\item{batch_corrected_df}{}
}
\value{

}
\description{
Title
}
