% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-panel.R
\name{create_panel}
\alias{create_panel}
\title{Create a CyTOF panel to be fill}
\usage{
create_panel(
  path,
  glob = "*.fcs",
  flowcore_params = list(truncate_max_range = FALSE),
  cytofkit2_params = list(mergeMethod = "all")
)
}
\arguments{
\item{path}{Path to search for cytometry files.}

\item{glob}{A wildcard aka globbing pattern (e.g. *.fcs) passed on
to \code{\link[base:grep]{base::grep()}} to filter paths.}

\item{flowcore_params}{Extra arguments passed to \code{\link[flowCore:read.flowSet]{flowCore::read.flowSet()}}.}

\item{cytofkit2_params}{Extra arguments passed to \code{\link[cytofkit2:cytof_exprsMerge]{cytofkit2::cytof_exprsMerge()}}.}
}
\value{
A tibble with 3 columns:
\enumerate{
\item \strong{fcs_colname}: Column filled it with channels found in data.
\item \strong{antigen}: Filled with NAs.
\item \strong{marker_class}: Filled with NAs.
}
}
\description{
Create a CyTOF panel to be fill
}
