% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyTOFHelpers-package.R
\docType{package}
\name{CyTOFHelpers-package}
\alias{CyTOFHelpers}
\alias{CyTOFHelpers-package}
\title{CyTOFHelpers: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
