% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-batch-correction-cydar.R
\name{plot_batch_corrected_results}
\alias{plot_batch_corrected_results}
\title{Plot before and after cydar batch correction}
\usage{
plot_batch_corrected_results(batch_corrected_df, cols = NULL, ...)
}
\arguments{
\item{output_file}{}
}
\value{

}
\description{
Plot before and after cydar batch correction
}
