% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-batch-correction-cydar.R
\name{prepare_batch_correction_format}
\alias{prepare_batch_correction_format}
\title{Prepare data to cydar batch correction format}
\usage{
prepare_batch_correction_format(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Prepare data to cydar batch correction format
}
