% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-batch-correction-cydar.R
\name{run_batch_correction}
\alias{run_batch_correction}
\title{Run cydar batch correction}
\usage{
run_batch_correction(batch_correction_df, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Run cydar batch correction
}
