% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-batch-correction-cydar.R
\name{save_batch_corrected_results}
\alias{save_batch_corrected_results}
\title{Save cydar batch corrected results}
\usage{
save_batch_corrected_results(
  batch_corrected_df,
  output_dir,
  prefix = "",
  suffix = ""
)
}
\arguments{
\item{suffix}{}
}
\value{

}
\description{
Save cydar batch corrected results
}
