% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-batch-correction-cydar.R
\name{split_by_batch_condition}
\alias{split_by_batch_condition}
\title{Split matrix by batch and condition}
\usage{
split_by_batch_condition(
  mat,
  metadata_sep = "_",
  batch_position = 1,
  condition_position = 2,
  control_condition_string = "Sp",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Split matrix by batch and condition
}
