% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing-panel.R
\name{write_panel}
\alias{write_panel}
\title{Write CyTOF panel}
\usage{
write_panel(panel, file, ...)
}
\arguments{
\item{panel}{A data frame or data frame extension (e.g. a tibble).}

\item{file}{Path to the output file.}

\item{...}{
  Arguments passed on to \code{\link[xlsx:write.xlsx]{xlsx::write.xlsx}}
  \describe{
    \item{\code{sheetName}}{a character string with the sheet name.}
    \item{\code{password}}{a String with the password.}
  }}
}
\description{
Write CyTOF panel
}
